refactor: allow install as a plugin app
Since the application could be installed as a plugin rather than a
standalone service, this commit refactors the code to be able to install
the install tha app as an edX Django Plugin App.

The plugin configuration is intentionally not placed inside the settings
module to avoid confusion and install of unrelated requirements like
environ or dotent dependencies.